feat: refactor dashboard view to use DashCards component for improved code organization and readability
