+++ components/dashboard/dashboarview.tsx
@@ -1,9 +1,6 @@
 'use client';
 
-import { Users, Shield, Truck, Clock, Package, Building } from 'lucide-react';
-import { Card } from '@/components/ui/card';
-
-import { JSX, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import {
   Order,
   TransformedOrder as OriginalTransformedOrder,
@@ -29,6 +26,7 @@ import { toast } from '@/hooks/use-toast';
 import { dashboardTotals, locations } from '@/types/dashboard';
 import { dashboardTotalsAPI, fetchLocations } from '@/actions/dashboard';
 import { v4 as uuidv4 } from 'uuid';
+import DashCards from './cards';
 
 export default function Dashboardview() {
   const [orders, setOrders] = useState<TransformedOrder[] | null>(null);
@@ -362,55 +360,8 @@ export default function Dashboardview() {
           <p className="text-slate-600">Here&apos;s Your Report</p>
         </div>
 
-        {/* Stats Row 1 */}
-        <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4">
-          <StatCard
-            title="Total Employees"
-            value={dashboarTotals.totalEmployees ?? 0}
-            icon={<Users className="h-6 w-6 text-white" />}
-            color="bg-blue-400"
-          />
-          <StatCard
-            title="Total Admins"
-            value={dashboarTotals.totalAdmins ?? 0}
-            icon={<Shield className="h-6 w-6 text-white" />}
-            color="bg-purple-400"
-          />
-          <StatCard
-            title="Delivered Items"
-            value={dashboarTotals.totalDelivered ?? 0}
-            icon={<Truck className="h-6 w-6 text-white" />}
-            color="bg-red-400"
-          />
-          <StatCard
-            title="Pending Items"
-            value={dashboarTotals.totalPending ?? 0}
-            icon={<Clock className="h-6 w-6 text-white" />}
-            color="bg-yellow-400"
-          />
-        </div>
+        <DashCards dashboarTotals={dashboarTotals!} locations={locations!} />
 
-        {/* Stats Row 2 */}
-        <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4">
-          <StatCard
-            title="Total Orders"
-            value={dashboarTotals.totalOrders ?? 0}
-            icon={<Package className="h-6 w-6 text-white" />}
-            color="bg-emerald-400"
-          />
-          <StatCard
-            title="Cities"
-            value={locations.length}
-            icon={<Building className="h-6 w-6 text-white" />}
-            color="bg-blue-400"
-          />
-          <StatCard
-            title="Picked Up Items"
-            value={dashboarTotals.totalPickedup ?? 0}
-            icon={<Users className="h-6 w-6 text-white" />}
-            color="bg-purple-400"
-          />
-        </div>
         {orders ? (
           <DataTable
             orderAmount={orderAmount}
@@ -466,26 +417,5 @@ export default function Dashboardview() {
         )}
       </div>
     </main>
-  );
-}
-
-type StatCard = {
-  title: string;
-  value: number;
-  color: string;
-  icon: JSX.Element;
-};
-
-function StatCard({ title, value, icon, color }: StatCard) {
-  return (
-    <Card className="p-6">
-      <div className="flex justify-between items-center">
-        <div>
-          <p className="text-sm text-gray-500">{title}</p>
-          <p className="text-4xl font-bold">{value}</p>
-        </div>
-        <div className={`rounded-full p-3 ${color}`}>{icon}</div>
-      </div>
-    </Card>
   );
 }
