+++ lib/utils.ts
@@ -59,3 +59,5 @@ export function priceCalculator(weight: number, basePrice: number): number {
     return basePrice! * weight;
   }
 }
+
+
