+++ types/StatCard.ts
@@ -1,0 +1,6 @@
+export type StatCard = {
+  title: string;
+  value: number;
+  color: string;
+  icon: React.ReactNode;
+};
